kafka-producer/requirements.txt
kafka-python==2.0.2
schedule==1.1.0
aiokafka==0.7.2
Faker==15.1.3
